Added DTO validation,updated Level and Review models,refined repositories,updated review controller

+++ src/main/java/com/example/unit2_backendproject_fordanceschool/repository/LevelRepository.java
@@ -4,6 +4,6 @@ import com.example.unit2_backendproject_fordanceschool.model.Level;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 //creating repository for level table helps for CRUD operations
-    public interface LevelRepository extends JpaRepository<Level, Long> {
+    public interface LevelRepository extends JpaRepository<Level, Integer> {
     }
